Added Grab to Objects
All four mission 2 objects now have the grab tag and can be picked up.
Whether or not they work with the goal is unknown. For some reason, the
city does not have collision as of the newest update. This is strange,
though assumed to be easily rectifiable if needed to be fixed.