Randomize the texture for meshes with multiple
Set it up in the blueprint Model_BoxOfFood. Whenever the blueprint is
dragged to the stage it adds the box of food with a randomly assigned
texture. Check the blueprint because it might be possible to not run the
top half every time but I cant figure out a way to set it up as a
blueprint.